refactor: reorganize GitHub blob loaders into document-loaders directory
- Move blob-download-loader.ts to document-loaders/blobs/archive-loader.ts
- Move blob-loader.ts to document-loaders/blobs/tree-loader.ts
- Update import paths in index.ts
- Better naming to distinguish archive (tarball) vs tree API approaches
